Remove client credentials interactor
